Update add-domain
The scammer created a sponsored ad on google so when you're in Mexico and search for "CFE" or "Comision Federal de Electricidad" you get served this ad.

This add looks just like that of the Mexican government owned electric company (it even shows the correct URL under the ad name), but upon clicking on the link it actually sends you to the domain: 

account-cfe.mx

Which has an exact replica of the power company's website, and it asks you to fill your user data to log in (which turns out to be a dead page because you can introduce garbage and will still let you in. But upon login (and before you can see your power bill) it asks you that for security you need to introduce credit card information which they steal.

The DNS domain is registered under a guy in Nevada, and not the mexican government and if I as an IT professional fell for this I'm pretty most regular people will too.

I urge you to add this page to your scammers list.